Return status of `lexer_collect_token`
Prior to this commit there was no way to know whether
`lexer_collect_token` had succeeded in collecting a valid token or not.
This commit resolves this issue by requiring that the result be passed
through a reference as a function argument and returning the status of
the operation as a boolean value.