Make all variables 64-bits in size
Because YASM infers the size of memory based on the size of registers
being read to/written from unless our variables are at least 64-bits in
size writing to a variable results in a memory overwrite. Because our
code generator currently only supports allocation and usage of 64-bit
registers this issue is being fixed by making all variables 64-bits in
size. This is of course an issue (albeit one of a much smaller scale)
that should be resolved later on--probably in the self-hosted compiler.